+++ src/realpaver/IntervalUnion.cpp
@@ -202,7 +202,7 @@ void IntervalUnion::contract(const Interval& x)
    v_[last] &= x;
 
    // removes the intervals after last
-   if (last < v_.size()-1)
+   if (last < (int)(v_.size()-1))
    {
       auto it = v_.begin();
       std::advance(it, last+1);
